time_complexity_analyzer/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.8
pytest>=6.2
pytest-django>=4.1
djangorestframework-simplejwt>=4.8
django-cors-headers>=3.7
numpy>=1.21.0
scipy>=1.7.0